feat: Index에서 인덱스가 1만큼 증가한다
