+++ README.md
@@ -89,5 +89,14 @@ Result클래스의 필요성 <- 우승 상품 검증에 관한 로직
 -[x] Line의 첫번째 지점에서 왼쪽으로 연결 됐을 때 예외가 발생한다
 -[x] Line의 마지막 지점에서 오른쪽으로 연결 됐을 때 예외가 발생한다
 -[x] Line의 선택된 점은 연결된 가로선에 따라 움직인다
+-[x] Index는 증가하면 인덱스가 1 증가한다
 Point가 서로 연결되어 있고
 Direction에 Index를 넘겨주면 그 다음 인덱스를 준다
+
+
+(boolean , boolean) 에서 Direction으로 넘어가니까 (ture, true)가 자연스럽게 없어짐
+
+사다리 게임은 각 라인들에게 역할을 넘겨주는 형태가 될 거고, 
+각 라인들은 해당위치에 존재하는 방향을 찾아서,
+방향한테 역할을 넘겨주는 형태를 갖게 될거고, 
+각 방향들은 자신의 위치에 맞게 움직이게 된다
