+++ src/main/java/domain/Direction.java
@@ -1,0 +1,10 @@
+//package domain;
+//
+//public enum Direction {
+//    LEFT, RIGHT, STRAIGHT
+//
+//    Direction(){
+//
+//    }
+//
+//}
